Directly import certificate to AWS certificate manager (#128)
## Problem
We create an AWS certificate, and then refer to it using a data resource within. the static buckets module. This works if it exists and on first creation. However, if the certificate has expired, the indirect reference breaks planning, and terraform can't create the certificate again.

## Solution
Directly refer the certificate to the static buckets. This explicitly shows Terraform the dependency, so it can create the certificate if it doesn't exist.