Don't destroy certificates that don't exist.
## Problem:
If a certificate expires, AWS will helpfully delete it. Then when rotating, Terraform also tries to delete it, but it doesn't exist. This blocks creation, and we loop.

## Solution
Desired behavior is to create the new certificate, and then try to delete the old one. If it isn't there, deletion will fail and so will the run. The next plan will notice the new certificate is correctly there, and the old one isn't, so it will leave it alone, breaking the cycle.